+++ src/pidcontrollers/level.hpp
@@ -1,7 +1,7 @@
 /*
    level.hpp : PID controller for Level mode
 
-   Copyright (c) 2018 Juan Gallostra and Simon D. Levy
+   Copyright (c) 2019 Simon D. Levy, Juan Gallostra Acin, Pep Marti-Saumell
 
    Author: Juan Gallostra
 
@@ -64,7 +64,7 @@ namespace hf {
             {
             }
 
-            bool modifyDemands(state_t & state, demands_t & demands, float currentTime)
+            bool modifyDemands(eskf_state_t & state, demands_t & demands, float currentTime)
             {
                 (void)currentTime;